+++ src/Pages/App.tsx
@@ -2,14 +2,14 @@ import localforage from "localforage";
 import { Route, Routes, useNavigate } from "react-router-dom";
 
 import CreateVFEForm from "../Pages/CreateVFE.tsx";
-import { VFE } from "../DataStructures.ts";
-import { load } from "../FileOperations.ts";
+import { VFE } from "./PageUtility/DataStructures.ts";
+import { load } from "./PageUtility/FileOperations.ts";
 import LandingPage from "../Pages/LandingPage.tsx";
 import PhotosphereEditor from "../Pages/PhotosphereEditor.tsx";
-import PhotosphereViewer from "../PhotosphereViewer.tsx";
+import PhotosphereViewer from "./PhotosphereViewer.tsx";
 import Prototype from "../Prototype/Prototype.tsx";
-import { convertRuntimeToStored, convertVFE } from "../VFEConversion.ts";
-import VFELoader from "../VFELoader.tsx";
+import { convertRuntimeToStored, convertVFE } from "./PageUtility/VFEConversion.ts";
+import VFELoader from "./PageUtility/VFELoader.tsx";
 
 // Main component acts as a main entry point for the application
 // Should decide what we are doing, going to LandingPage/Rendering VFE
