+++ src/Pages/CreateVFE.tsx
@@ -9,9 +9,9 @@ import {
   VFE,
   calculateImageDimensions,
   newID,
-} from "../DataStructures.ts";
+} from "./PageUtility/DataStructures.ts";
 import Header, { HeaderProps } from "../UI/Header.tsx";
-import PhotosphereLocationSelector from "../PhotosphereLocationSelector.tsx";
+import PhotosphereLocationSelector from "./PhotosphereLocationSelector.tsx";
 import { alertMUI } from "../UI/StyledDialogWrapper.tsx";
 
 //import { PhotosphereCenterFieldset } from "./buttons/AddPhotosphere.tsx";
