+++ src/Pages/PhotosphereSelector.tsx
