+++ src/buttons/AddPhotosphere.tsx
@@ -12,8 +12,8 @@ import {
   TextField,
 } from "@mui/material";
 
-import { Photosphere, VFE, newID } from "../DataStructures";
-import PhotosphereLocationSelector from "../PhotosphereLocationSelector.tsx";
+import { Photosphere, VFE, newID } from "../Pages/PageUtility/DataStructures.ts";
+import PhotosphereLocationSelector from "../Pages/PhotosphereLocationSelector.tsx";
 import { alertMUI } from "../UI/StyledDialogWrapper.tsx";
 
 interface AddPhotosphereProps {
