+++ src/buttons/RemovePhotosphere.tsx
@@ -11,10 +11,10 @@ import {
   Typography,
 } from "@mui/material";
 
-import { VFE } from "../DataStructures";
+import { VFE } from "../Pages/PageUtility/DataStructures";
 import PhotosphereSelector, {
   PhotosphereSelectorProps,
-} from "../PhotosphereSelector";
+} from "../Pages/PhotosphereSelector";
 import { alertMUI } from "../UI/StyledDialogWrapper";
 
 interface RemovePhotosphereProps {
